refactoring udany
