--- accounts/urls.py
@@ -1,8 +1,0 @@
-from django.urls import path
-#from .views import SignUpView
-from .views import register
-
-urlpatterns = [
-    #path('signup/', SignUpView.as_view(), name='signup'),
-    path('register/', register, name='register'),
-]
